Refactor DB modules into shared package
