+++ ppat_db/__init__.py
@@ -1,0 +1,1 @@
+"""PPAT 공통 데이터베이스 모듈."""
